Puppet 4 Support and Testing
This is a refactor of the entire module to test against Puppet 4 with
strict variables set.

The structure of the module has been changed to ensure that users can
use 'include' on the subclasses instead of relying on the main 'haveged'
class as a profile.

Acceptance tests were added for EL6 and EL7 and a bug in the use of the
package provider with ensure 'purged' was addressed.

The original spirit of the module was preserved so that the entire setup
should be 100% backward compatible.